+++ clang/lib/Sema/SemaAvailability.cpp
@@ -452,28 +452,6 @@ static void DoEmitAvailabilityWarning(Sema &S, AvailabilityResult K,
       }
     }
   }
-  // Suppress -Wdeprecated-declarations in purely implicit special-member functions.
-  if (K == AR_Deprecated) {
-    if (const FunctionDecl *FD = S.getCurFunctionDecl()) {
-      // Implicit, defaulted constructor (default / copy / move)
-      if (auto *CD = dyn_cast<CXXConstructorDecl>(FD)) {
-        if (CD->isImplicit() && CD->isDefaulted())
-          return;
-      }
-      //Implicit, defaulted destructor
-      if (auto *DD = dyn_cast<CXXDestructorDecl>(FD)) {
-        if (DD->isImplicit() && DD->isDefaulted())
-          return;
-      }
-      //Implicit copy- or move-assignment operator
-      if (auto *MD = dyn_cast<CXXMethodDecl>(FD)) {
-        if (MD->isImplicit() &&
-            (MD->isCopyAssignmentOperator() || MD->isMoveAssignmentOperator()))
-          return;
-      }
-    }
-  }
-
 
   switch (K) {
   case AR_NotYetIntroduced: {
@@ -569,6 +547,14 @@ static void DoEmitAvailabilityWarning(Sema &S, AvailabilityResult K,
     return;
   }
   case AR_Deprecated:
+    // Suppress -Wdeprecated-declarations in purely implicit special-member
+    // functions.
+    if (auto *MD = dyn_cast_if_present<CXXMethodDecl>(S.getCurFunctionDecl());
+        MD && MD->isImplicit() && MD->isDefaulted() &&
+        (isa<CXXConstructorDecl, CXXDestructorDecl>(MD) ||
+          MD->isCopyAssignmentOperator() || MD->isMoveAssignmentOperator())) {
+       return;
+    }
     if (ObjCPropertyAccess)
       diag = diag::warn_property_method_deprecated;
     else if (S.currentEvaluationContext().IsCaseExpr)